внесла изменения, перенесла в файд data.js
